+++ lang/crates/vo-codegen/src/expr/selector.rs
@@ -225,7 +225,7 @@ fn compile_pkg_qualified_name(
             if let Some(global_idx) = ctx.get_global_index(obj_key) {
                 let type_key = info.expr_type(expr.id);
                 let slots = info.type_slot_count(type_key);
-                func.emit_global_get(dst, global_idx as u16, slots);
+                func.emit_global_get(dst, global_idx, slots);
                 Ok(())
             } else {
                 Err(CodegenError::Internal(format!(
